Fix all cases of error shadowing
Use "err" if we are in a if-clause, if not start with e and then e1, e2,
etc. This to prevent things for getting too unwieldly, something more
elegant might be needed in the long run. There is also a vague plan to
autogenerate this - problem there is the weirdness some RRs for parsing
(depending on previous values, etc. etc.)

Signed-off-by: Miek Gieben <miek@miek.nl>